RF-11904 add missing query string arguments
`include_feedback` is required to receive the failed test details in
the response.
`skip_mark_as_viewed` is required to not mark the test result as viewed.
This was problematic before, because marking a test as viewed also
changed the `updated_at` value for that test result, which we use
for calculating test duration.