RF-20682 Remove travis config
The build statuses from Travis don't appear to be getting back to
Github, and I'm not sure why we're building on both Travis and CircleCI.
Plus Travis are deprecating travis-ci.org and I don't want to do the
work to move over.

So remove Travis as a CI/CD platform.